Make search in selected files opt-in
By default we search everywhere now. To search only on the selected files,
start your query with a ':' (a single colon sign).

The colon sign will be removed from the query but will enable the selected
files filter.

I did this because I prefer the default to be search everywhere.
I rarely use the "search in selected files" feature.

Signed-off-by: Pedro Melo <melo@simplicidade.org>